# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Tasks API",
        "description": "User Tasks API",
        "contact": {
            "email": "evaldo@klocktecnologia.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth": {},
        "/api/tasks": {
            "get": {
                "summary": "Retrieves a collection of tasks for the authenticated user.",
                "operationId": "faf0fc05c6a1b0c6bfe6eadb833f2f64",
                "responses": {
                    "200": {
                        "description": "The collection of tasks."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "summary": "Creates a new task for the authenticated user.",
                "operationId": "2faae07edd85c912c06b857cfe0a5af9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "I need to buy milk for breakfast."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully."
                    },
                    "400": {
                        "description": "Invalid request data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "summary": "Register a new user.",
                "operationId": "4d4c8281f9b6a7269a227a4c55758908",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully."
                    },
                    "400": {
                        "description": "Invalid request data."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "Log in a user.",
                "operationId": "9053332bb6088000c55bfe467b61ab5a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token generated successfully."
                    },
                    "401": {
                        "description": "Invalid credentials."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/tasks/{taskId}": {
            "put": {
                "summary": "Updates a task.",
                "operationId": "341b297adc3b319e5b373c3dab4a4d21",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "The ID of the task to be updated.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "I need to buy milk for breakfast."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task updated successfully."
                    },
                    "404": {
                        "description": "Task not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "summary": "Deletes a task.",
                "operationId": "66903d145e9d9a4fa9e9263ab7c5bb16",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "The ID of the task to be deleted.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Task deleted successfully."
                    },
                    "404": {
                        "description": "Task not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}